Refactor
- Updated comments by functions
- Added comments by functions
- Combined CSS files
- Removed useless files
- Improved code quality
