+++ app/routes/image.js
@@ -1,0 +1,12 @@
+// app/routes/image.js
+
+/**
+ * Expose image routes
+ */
+module.exports = function(app) {
+
+    app.post('/upload', function (req, res) {
+        res.json(req.files.image);
+    });
+    
+};
