--- public/CSS/home-page.css
@@ -1,104 +1,0 @@
-.img-responsive {
-    display: block;
-    margin: 0 auto;
-    max-height: 903px;
-    border-bottom: 1px solid black;
-}
-
-#Carousel{
-    margin-top: -20px!important;
-    margin-bottom: 20px;
-    background-color: #252525;
-    height:903px;
-}
-
-.carousel-indicators {
-    background: -moz-linear-gradient(left,  rgba(0,0,0,0) 0%, rgba(0,0,0,0.65) 50%, rgba(0,0,0,0) 100%); /* FF3.6+ */
-    background: -webkit-gradient(linear, left top, right top, color-stop(0%,rgba(0,0,0,0)), color-stop(50%,rgba(0,0,0,0.65)), color-stop(100%,rgba(0,0,0,0))); /* Chrome,Safari4+ */
-    background: -webkit-linear-gradient(left,  rgba(0,0,0,0) 0%,rgba(0,0,0,0.65) 50%,rgba(0,0,0,0) 100%); /* Chrome10+,Safari5.1+ */
-    background: -o-linear-gradient(left,  rgba(0,0,0,0) 0%,rgba(0,0,0,0.65) 50%,rgba(0,0,0,0) 100%); /* Opera 11.10+ */
-    background: -ms-linear-gradient(left,  rgba(0,0,0,0) 0%,rgba(0,0,0,0.65) 50%,rgba(0,0,0,0) 100%); /* IE10+ */
-    background: linear-gradient(to right,  rgba(0,0,0,0) 0%,rgba(0,0,0,0.65) 50%,rgba(0,0,0,0) 100%); /* W3C */
-    filter: progid:DXImageTransform.Microsoft.gradient( startColorstr='#00000000', endColorstr='#00000000',GradientType=1 ); /* IE6-9 */
-}
-
-.page-info-carousel {
-    position: absolute;
-    height:200px;
-    width:40%;
-    color:white;
-    top:70%;
-    word-wrap: break-word;
-    text-align: center;
-    margin: 0 auto;
-    left: 0;
-    right: 0;
-    overflow: hidden;
-    transition: background-color 0.5s ease;
-    background-color: rgba(25,25,25,0.3);
-}
-
-.page-info-carousel:hover {
-    background-color: rgba(25,25,25,0.7);
-}
-
-
-/*
-Student work.
-*/
-
-.unplaced {
-    height:300px;
-    margin-bottom: 30px;
-}
-
-.unplaced:hover {
-
-}
-
-.page-info{
-    position: absolute;
-    height:200px;
-    width:200px;
-    border-radius: 50%;
-    background-color: rgba(255,255,255,0.8);
-    border: 5px solid white;
-    left:calc(50% - 100px);
-    top:calc(50% - 110px);
-    padding-top: 80px;
-    word-wrap: break-word;
-    text-align: center;
-}
-
-.page-info > h3 {
-    color:#252525;
-    text-align: center;
-}
-
-.view-more {
-    display: table;
-    margin: 0 auto;
-    margin-top: 50px;
-}
-
-#slider1_container, #slides {
-    height:903px;
-    width:1080px;
-}
-
-/*
-votes
- */
-
-.visibility-of-votes {
-    visibility:hidden;
-    opacity:0;
-    transition:visibility 0s linear 0.5s,opacity 0.5s linear;
-    font-size: 2em;
-}
-
-.unplaced:hover .visibility-of-votes{
-    visibility:visible;
-    opacity:1;
-    transition-delay:0s;
-}
