--- public/CSS/mainCustom.css
@@ -1,389 +1,0 @@
-body {
-    margin-bottom: 200px;
-    position: relative;
-}
-/**
- * Basic styles
- */
-img {
-    max-width: 100%;
-    height: auto;
-}
-
-/**
- * Navigation
- */
-
-.navbar-brand img {
-    height: 40px;
-    margin-top: -15px;
-    margin-bottom: -15px;
-}
-
-/**
- * Slide out menu
- */
-.menu, .menu-right {
-    overflow-x: hidden;
-    position: relative;
-    left: 0;
-    right: 0;
-}
-
-.menu-open {
-    right: auto;
-    left: 380px;
-}
-
-.menu-open-right {
-    left: auto;
-    right: 380px;
-}
-
-.menu-open .menu-side {
-    left: 0;
-    padding-top: 60px;
-}
-
-.menu-open-right .menu-side-right {
-    right: 0;
-    padding-top: 60px;
-}
-
-.menu-side-right, .menu-side, .menu {
-    -webkit-transition: 0.3s ease;
-    -moz-transition: 0.3s ease;
-    -o-transition: 0.3s ease;
-    -ms-transition: 0.3s ease;
-    transition: 0.3s ease;
-}
-
-.menu-side {
-    background: #333;
-    color: #fff;
-    position: fixed;
-    top: 0;
-    left: -380px;
-    width: 380px;
-    padding: 10px;
-    height: 100%;
-}
-
-.menu-side-right {
-    background: #333;
-    color: #fff;
-    position: fixed;
-    top: 0;
-    right: -380px;
-    width: 380px;
-    padding: 10px;
-    height: 100%;
-}
-
-.project-actions {
-    position: relative;
-    text-align: center;
-}
-
-.menu-toggle{
-    position: absolute;
-    top: 0;
-    left: 0;
-}
-
-.menu-toggle-right{
-    position: absolute;
-    top: 0;
-    right: 0;  
-}
-
-/**
- * Layout picks
- */
-.project-layouts {
-    margin-top: 40px;
-}
-
-.layout-item {
-    width: 100%;
-    text-align: center;
-}
-
-.layout-item i {
-    font-size: 90px;
-    margin: 0 10px;
-}
-
-.layout {
-    position: relative;
-}
-
-.remove-layout {
-    display: none;
-    position: absolute;
-    top: 0;
-    right: 10px;
-    font-size: 24px;
-    width: 24px;
-    height: 24px;
-}
-
-.layout:hover .remove-layout {
-    display: block;
-}
-
-.layout-image {
-    padding: 20px 0;
-    text-align: center;
-}
-
-.layout-image img {
-    max-width: 100%;
-    height: auto;
-}
-
-.layout-image-text {
-    padding: 20px 0;
-}
-
-.layout-image-text .image {
-    text-align: center;
-}
-
-.layout-image-text .image img {
-    max-width: 100%;
-    height: auto;
-}
-
-.layout-text-image {
-    padding: 20px 0;
-}
-
-.layout-text-image .image {
-    text-align: center;
-}
-
-.layout-text-image .image img {
-    max-width: 100%;
-    height: auto;
-}
-
-.layout-video {
-    padding: 20px 0;
-}
-
-.layout-video .embed-container { 
-    position: relative; 
-    padding-bottom: 56.25%; 
-    height: 0; 
-    overflow: hidden; 
-    max-width: 100%; 
-    height: auto; 
-} 
-
-.layout-video .embed-container iframe, 
-.layout-video .embed-container object, 
-.layout-video .embed-container embed { 
-    position: absolute; 
-    top: 0; 
-    left: 0; 
-    width: 100%; 
-    height: 100%; 
-}
-
-.modal-lg {
-    width: 900px;
-}
-
-.froala-editor.f-basic.f-scroll-abs {
-    position: initial;
-}
-
-.select-text,
-.select-image {
-    text-align: center;
-}
-
-.select-text i,
-.select-image i {
-    font-size: 90px;
-    margin-right: 10px;
-}
-
-/**
- * Project page
- */
-.header-image {
-    position: relative;
-    margin-bottom: 100px;
-}
-
-.header-image > img {
-    width: 100%;
-    height: auto;
-}
-
-.header-image .project-logo {
-    border-radius: 50%;
-    border: 5px solid white;
-    position: relative;
-}
-
-.header-image .project-logo img {
-    position: absolute;
-    bottom: -85px;
-    left: calc(50% - 105px);
-}
-
-/*About page
-*/
-.round-logo-background {
-    position: absolute;
-    left: calc(50% - 100px);
-    margin-top: -120px;
-    background-image: url(http://lorempixel.com/500/300);
-    background-position: 50% 50%;
-    background-repeat: no-repeat;
-    border-radius: 50%;
-    width: 200px;
-    height: 200px;
-}
-
-.title {
-    margin-top:60px; 
-}
-
-.inner-image img {
-    margin-top: 145px; 
-    left: calc(50% - 100px);
-    top: calc(100% - 135px);
-}
-
-.inner-text  {
-    margin-top: 50px;
-    padding-right: 125px;
-    padding-left: 125px;
-}
-
-.about-container {
-    margin-top:150px; 
-    border: 1px solid #e7e7e7;
-}
-/**
- * Other stuff
- */
-
-.margin-view {
-    margin-top: 70px;
-}
-
-.firstPlace {
-    border: 1px solid black;
-    height: 800px!important;
-    background-color: #428bca;
-}
-
-.secondAndThirdPlace {
-    border: 1px solid black;
-    height: 400px!important;
-    background-color: #5cb85c;
-}
-
-.unplaced {
-    border: 1px solid black;
-    height: 200px;  
-    background-color: #d9534f;
-}
-
-.full-row {
-    margin: 0 auto; 
-}
-
-.imagebanner {
-    margin: 0 auto;
-    height:200px;
-    background-color: #428bca;
-}
-
-
-.footer {
-    color: white;
-    background: #252525;
-    
-    height: 200px;
-    left: 0;
-    bottom: -200px;
-    width: 100%;
-}
-
-.footer h3{
-    text-align: center;
-}
-
-.footer a {
- color: white;   
-}
-
-.team-image img {
-    border-radius: 50%;
-    display: inline-block;
-    background-color: green;
-    width:200px;
-    height:200px;
-    display: inline-block;
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-    margin: auto;
-    margin-top:350px!important;    
-}
-
-.page-header {
-    border: none;
-}
-
-.header-image {
-    min-height: 250px;
-    margin-bottom: 0;
-}
-
-.select-text,
-.select-image,
-.select-youtube {
-    text-align: center;
-}
-
-.select-text i,
-.select-image i,
-.select-youtube i {
-    font-size: 90px;
-    margin-right: 10px;
-}
-
-.project-title-edit {
-    text-align: center;
-    font-size: 32px;
-}
-
-.page-header {
-    margin: 40px auto 100px auto;
-}
-
-.projectsemesterform {
-    position: absolute;
-    top: 0;
-    right: 0;
-}
-
-#createdcontent .remove-layout,
-#createdcontent .projectsemesterform {
-    display: none;
-}
-
-.tab-pane {
-    padding: 10px;
-    border-left: 1px solid #ddd;
-    border-right: 1px solid #ddd;
-}
