--- public/CSS/student-work-page.css
@@ -1,25 +1,0 @@
-.slide-student {
-    position: absolute!important;
-    top:10%;
-    left:0;
-    -webkit-transition:all ease 1.5s;
-    transition:all ease 1.5s;
-    height:300px;
-    width:100%;
-}
-
-.slide-student:nth-child(2) {
-    left:0px;
-}
-
-.slide-student:nth-child(3) {
-    left:calc(1920px + 1920px/2);
-}
-
-#left-student-slide, #right-student-slide {
-    margin-top:-20px;
-}
-
-#controles{
-    visibility: visible;
-}
