+++ public/controllers/home.js
@@ -1,6 +1,12 @@
+// public/controllers/home.js
+
+/**
+ * Expose homepage controller and home route
+ */
 (function() {
     
     angular.module('home', [])
+
         .config(['$routeProvider', '$locationProvider', function($routeProvider, $locationProvider) {
             $routeProvider    
                 .when('/', {
@@ -14,6 +20,9 @@
             $locationProvider.html5Mode({ enabled: true, requireBase: false });
         }])
 
+        /**
+         * Show the homepage with some projects.
+         */
         .controller('HomePageController', ['$http', '$scope', '$routeParams', function($http, $scope, $routeParams) {
             $http.get('/api/projects/')
                 .success (function (data) {
