+++ public/controllers/login.js
@@ -4,33 +4,26 @@
  * Expose login controllers and routes
  */
 (function() {
+
     angular.module('login', [])
+        
         .config(['$routeProvider', '$locationProvider', function($routeProvider, $locationProvider) {
             $routeProvider    
                 .when('/signup', {
                     templateUrl: '/views/signup.html',
-                    controller: 'SignupController',
-                    controllerAs: 'signup',
                     caseInsensitiveMatch: true
                 })
                 .when('/login', {
                     templateUrl: '/views/login.html',
-                    controller: 'LoginController',
-                    controllerAs: 'login',
                     caseInsensitiveMatch: true
                 });
 
             $locationProvider.html5Mode({ enabled: true, requireBase: false });
         }])
-        .controller('LoginController', ['$http', '$scope', function($http, $scope) {
-            // Custom Login functionality
-        }])
-        .controller('IndexController', ['$http', '$scope', function($http, $scope) {
-            // Custom Index functionality
-        }])
-        .controller('SignupController', ['$http', '$scope', function($http, $scope) {
-            // Custom Signup functionality
-        }])
+
+        /**
+         * Provide a function so the user can post the form to log themself in.
+         */
         .controller('LoginForm', ['$http', '$scope', '$rootScope', function($http, $scope, $rootScope) {
             $scope.login = function() {
                 $http.post('/login', { 
@@ -41,6 +34,10 @@
                 $rootScope.loggedin = true;
             }
         }])
+
+        /**
+         * Provide a list of Profiles for the user to choose from when registering.
+         */
         .controller('SignupForm', ['$http', '$scope', function($http, $scope) {
             $http.get('/api/profiles')
                 .success(function (data) {
