--- public/css/customBob.css
@@ -1,119 +1,0 @@
-.profileinformation {
-    margin-top: 15px;
-}
-
-.profileinformation .form-margin {
-    margin-top: 20px;
-}
-
-.profileinformation .form-margin:last-child {
-    margin-bottom: 0px;
-}
-
-.project-logo-background {
-    position: absolute;
-    left: calc(50% - 100px);
-    background-size: cover;
-    background-repeat: no-repeat;
-    border-radius: 50%;
-    width: 200px;
-    height: 200px;
-}
-
-.project-logo .select-image {
-    width: 200px;
-    height: 200px;
-    margin: -110px auto 0 auto;
-}
-
-#vote {
-    display: none;
-}
-
-#createdcontent #vote {
-    display: block;
-    margin: 0 auto;
-    font-size: 40px;
-    width: 40px;
-}
-
-.layout-video .actions,
-#createdcontent .actions {
-    display: none;
-}
-
-.layout-video:hover .actions {
-    display: block;
-}
-
-.embed-container .actions {
-    z-index: 999;
-    position: absolute;
-    top: calc(50% - 75px);
-    left: calc(50% - 160px);
-}
-
-.embed-container .actions a {
-    display: inline-block;
-    text-align: center;
-    background: #eeeeee;
-    width: 150px;
-    height: 150px;
-    font-size: 100px;
-    margin-right: 10px;
-    line-height: 150px;
-}
-
-.embed-container .actions a:last-child {
-    margin-right: 0;
-    margin-left: 10px;
-}
-
-.page-header {
-    margin-top: 20px;
-    margin-bottom: 0;
-}
-
-.banner-background {
-    background: #eee;
-    padding: 50px 0 110px 0;
-}
-
-.profile-background,
-.logo-background {
-    background: #ddd;
-    padding: 50px 0;  
-    margin: 0 auto;  
-    width: 200px;
-    height: 200px;
-    border-radius: 50%;
-}
-
-.logo-background {
-    margin: -70px auto 0 auto;
-}
-
-.image-background,
-.text-background,
-.youtube-background {
-    padding: 40px;
-    background: #eee;
-}
-
-.select-text i, 
-.select-image i, 
-.select-youtube i {
-    font-size: 100px;
-    margin-right: 0;
-}
-
-.error-title {
-    margin: 20px auto 0 auto;
-    display: none;
-}
-
-.error-semester {
-    margin: 20px auto 0 auto;
-    display: none;   
-}
-
