+++ public/js/script.js
