+++ public/views/index.html
@@ -10,31 +10,26 @@
         <link rel="stylesheet" href="../css/froala_editor.min.css">
         <link rel="stylesheet" href="../css/froala_style.min.css">
         <link rel="stylesheet" href="../css/froala_content.min.css">
-        <link rel="stylesheet" href="../css/mainCustom.css">
-        <link rel="stylesheet" href="../css/home-page.css">
-        <link rel="stylesheet" href="../css/student-work-page.css">
-        <link rel="stylesheet" href="../css/customBob.css">
+        <link rel="stylesheet" href="../css/styles.css">
 
         <script src="//ajax.googleapis.com/ajax/libs/angularjs/1.3.6/angular.min.js"></script>
         <script src="//ajax.googleapis.com/ajax/libs/angularjs/1.3.6/angular-route.min.js"></script>
         
+        <script src="/modules/core.js"></script>    
+        <script src="/modules/growl.js"></script>
         <script src="/factories/httpInterceptor.js"></script>
-        <script src="/directives/common.js"></script>
-        <script src="/controllers/collection.js"></script>
+        <script src="/controllers/admin.js"></script>
+        <script src="/controllers/footer.js"></script>
         <script src="/controllers/home.js"></script>
+        <script src="/controllers/login.js"></script>
         <script src="/controllers/pages.js"></script>
-        <script src="/controllers/test.js"></script>
+        <script src="/controllers/profile.js"></script>
         <script src="/controllers/profiles.js"></script>
         <script src="/controllers/project.js"></script>
-        <script src="/controllers/footer.js"></script>
-	    <script src="/controllers/profile.js"></script>
-	    <script src="/controllers/login.js"></script>
-        <script src="/controllers/admin.js"></script>
-        <script src="/modules/growl.js"></script>
-        <script src="/modules/core.js"></script>
+        <script src="/directives/common.js"></script>
     </head>    
 
-    <body class="menu menu-right">
+    <body class="menu">
 
         <nav class="navbar navbar-default navbar-fixed-top" role="navigation">
             <div class="container-fluid">          
@@ -107,13 +102,13 @@
         <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.1/js/bootstrap.min.js"></script>
         <script src="../js/jquery.cookie.js"></script>
         <script src="../js/froala_editor.min.js"></script>
-        <script src="../js/plugins/block_styles.min.js"></script>
-        <script src="../js/plugins/colors.min.js"></script>
-        <script src="../js/plugins/font_family.min.js"></script>
-        <script src="../js/plugins/font_size.min.js"></script>
-        <script src="../js/plugins/lists.min.js"></script>
-        <script src="../js/plugins/tables.min.js"></script>
-        <script src="../js/fly-out-menu.js"></script>
+        <script src="../js/froala/block_styles.min.js"></script>
+        <script src="../js/froala/colors.min.js"></script>
+        <script src="../js/froala/font_family.min.js"></script>
+        <script src="../js/froala/font_size.min.js"></script>
+        <script src="../js/froala/lists.min.js"></script>
+        <script src="../js/froala/tables.min.js"></script>
+        <script src="../js/script.js"></script>
         <script src="../js/clamp.js"></script>
         <script src="../js/home-clamp.js"></script>
         <script src="../js/student-work-slider.js"></script>
